// front/src/app/core/modules/openapi/model/new-reply.ts
/**
 * MDD OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Author } from './author';


export interface NewReply { 
    message: string;
    author: Author;
}